Add support for multiple services in VCAP_SERVICES
Adds builder classes for creating clients suitable for Bluemix and non-Bluemix
envrionments. The Bluemix builder allows for matching against the 'name' and
'label' fields present in VCAP_SERVICES.

This supports situations where multiple instances of the MQ Light service are
bound to an application - or where the service label is not 'mqlight'.